SCJ-123: Rename component files
