+++ app/ClientSrc/vue/ApplicationTypeSelect/ApplicationTypeSelect.vue
