+++ app/ClientSrc/vue/ApplicationTypeSelect/index.js
