+++ app/Views/CoaBooking/_Layout.cshtml
@@ -27,12 +27,12 @@
 
 @section Scripts {
     <script src="/scjob/js/coa.js" asp-append-version="true"></script>
-    <script src="/scjob/dist/vue/coa.js" asp-append-version="true"></script>
+    <script src="/scjob/dist/vue/ApplicationTypeSelect.js" asp-append-version="true"></script>
     @RenderSection("Scripts", false)
 }
 
 @section Styles {
     @RenderSection("Styles", false)
     <link rel="stylesheet" href="/scjob/dist/css/coa.css" asp-append-version="true"/>
-    <link rel="stylesheet" href="/scjob/dist/vue/coa.js.css" asp-append-version="true"/>
+    <link rel="stylesheet" href="/scjob/dist/vue/ApplicationTypeSelect.js.css" asp-append-version="true"/>
 }
