+++ app/Views/ScBooking/AvailableTimes.cshtml
@@ -36,11 +36,12 @@
             </div>
         </div>
 
-        <div id="VueAvailableTimes" class="availableTimesContainer">
-            <availabletimes :location-id="@Model.BookingRegistryId"
-                            :available-dates="@availableDates.ToString()"
-                            :hearing-type="@Model.HearingTypeId">
-            </availabletimes>
+        <div id="VueHearingTimeSelect" class="availableTimesContainer">
+            <hearing-time-select
+                :location-id="@Model.BookingRegistryId"
+                :available-dates="@availableDates.ToString()"
+                :hearing-type="@Model.HearingTypeId"
+            ></hearing-time-select>
         </div>
 
         @for (int i = 0; i < Model.CourtFiles?.Length; i++)
@@ -86,7 +87,7 @@
 
 @section Scripts {
     <script src="/scjob/dist/vendor.js" asp-append-version="true"></script>
-    <script src="/scjob/dist/vue/booking-casesearch.js" asp-append-version="true"></script>
+    <script src="/scjob/dist/vue/HearingTimeSelect.js" asp-append-version="true"></script>
     <script src="/scjob/dist/lib/bootstrap-datepicker.js" asp-append-version="true"></script>
     <script type="text/javascript">
         $('#datepicker').datepicker({
